Правки настроек сервера
