--- packages/db/index.ts
@@ -1,13 +1,0 @@
-import { Sequelize, SequelizeOptions } from 'sequelize-typescript'
-
-const sequelizeOptions: SequelizeOptions = {
-  host: 'localhost',
-  port: Number(process.env.POSTGRES_PORT),
-  username: process.env.POSTGRES_USER,
-  password: process.env.POSTGRES_PASSWORD,
-  database: process.env.POSTGRES_DB,
-  dialect: 'postgres',
-}
-
-const sequelize = new Sequelize(sequelizeOptions)
-export default sequelize
